\name{compareTranscriptAndOutput}
\alias{compareTranscriptAndOutput}
\title{ Compare desired and actual output from running transcript-style tests.}
\description{
  Compare desired and actual output from running transcript-style
  tests.  This function is called by the scripts supplied with the
  EasyTests package.
}
\usage{
compareTranscriptAndOutput(name, tests, results, verbose = TRUE)
}
\arguments{
  \item{name}{ Name of the file of tests. }
  \item{tests}{ A list of test blocks, as generated by \code{parseTranscriptFile} }
  \item{results}{ A list of test blocks, as generated by \code{parseTranscriptFile} }
  \item{verbose}{ A logical flag indicating whether to write progress reports to stdout. }
}
\details{
  Compares tests line-by-line, with some control over how matching is
  done and what mismatches counts as errors, warnings, or informational events.
}
\value{
  Returns an object of class \code{RtTestSetResults}, for which there
  are summary and print methods.
}
\author{ Tony Plate }
\keyword{ misc }
