\name{plus}
\alias{plus}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ addition }
\description{
  Add two numbers.  This function is created for testing purposes.
}
\usage{
plus(x, y)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a number }
  \item{y}{ a number }
}
\value{
  The result of adding \code{x} and \code{y}
}
\keyword{ misc }
